requirements_temp.txt
chromadb==1.0.20
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
numpy==2.3.1
pandas==2.3.0
python-dotenv==1.1.1
rank-bm25==0.2.2
sentence-transformers==5.1.0
streamlit==1.46.1
streamlit-audiorecorder==0.0.6
torch==2.8.0
transformers==4.55.4
